+++ Includes/RewriteRules.php
@@ -2,6 +2,11 @@
 
 namespace VOHTMLSitemap\Includes;
 
+use VOHTMLSitemap\Includes\Pages\LastWeek;
+use VOHTMLSitemap\Includes\Pages\ThisWeek;
+use VOHTMLSitemap\Includes\Pages\Today;
+use VOHTMLSitemap\Includes\Pages\Yesterday;
+
 class RewriteRules
 {
     public static function init(): void
@@ -25,7 +30,7 @@ class RewriteRules
         }
 
         add_rewrite_rule(
-            "^{$page->post_name}\/([0-9]{4})\/([0-9]{1,2})\/([0-9]{1,2})$",
+            "^{$page->post_name}\/([0-9]{4})\/?([0-9]{1,2})\/([0-9]{1,2})$",
             'index.php?page_id=' . $page->ID . '&vo-html-sitemap=true&vo-html-sitemap-year=$matches[1]&vo-html-sitemap-month=$matches[2]&vo-html-sitemap-day=$matches[3]',
             'top'
         );
@@ -41,6 +46,16 @@ class RewriteRules
             'top'
         );
 
+        $sitemapDates = [new Today(), new Yesterday(), new ThisWeek(), new LastWeek()];
+
+        foreach ($sitemapDates as $date) {
+            add_rewrite_rule(
+                "{$page->post_name}/{$date->getSlug()}$",
+                'index.php?page_id=' . $page->ID . '&vo-html-sitemap=true&vo-html-sitemap-range=' . $date->getSlug(),
+                'top'
+            );
+        }
+
         add_rewrite_rule(
             "{$page->post_name}$",
             'index.php?page_id=' . $page->ID . '&vo-html-sitemap=true',
@@ -51,6 +66,7 @@ class RewriteRules
     public static function addQueryVars(array $vars): array
     {
         $vars[] = 'vo-html-sitemap';
+        $vars[] = 'vo-html-sitemap-range';
         $vars[] = 'vo-html-sitemap-year';
         $vars[] = 'vo-html-sitemap-month';
         $vars[] = 'vo-html-sitemap-day';