+++ vo-html-sitemap.php
@@ -5,7 +5,7 @@
  * Description:       Van Ons plugin to add a HTML sitemap to your site.
  * Author:            Van Ons
  * Author URI:        https://van-ons.nl/
- * Version:           1.0.1
+ * Version:           1.0.2
  * Text Domain:       vo-html-sitemap
  * Domain Path:       /languages
  * Requires at least: 6.4
@@ -26,7 +26,7 @@ if( !defined('ABSPATH') || !function_exists('add_filter') ) {
 spl_autoload_register(__NAMESPACE__ . '\\autoload');
 
 // define plugin constants
-define('VOHTMLSITEMAP_VERSION', '1.0.1');
+define('VOHTMLSITEMAP_VERSION', '1.0.2');
 
 define('VOHTMLSITEMAP_ROOT', dirname(__FILE__) . '/');
 define('VOHTMLSITEMAP_ROOT_FILE', __FILE__);