+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All changes to this project will be documented in this file.
 
+## 1.0.2 - 2024-08-19
+
+- Refactor sitemap items to pages
+- Added range sitemaps: Today, Yesterday, This week, Last week.
+
+## 1.0.1 - 2024-08-19
+
+- Post type bugfix
+- Use grid instead of lists for sitemaps except for the main sitemap.
+
 ## 1.0.0 - 2024-08-19
 
 - initial release